+++ NEWS.md
@@ -26,6 +26,14 @@
 
 * Fix issues of `TimeArray` column names copying. (#418)
 
+* `ta[]` throws `BoundsError` now. (#420)
+
+  ```julia
+  julia> cl[]
+  ERROR: BoundsError: attempt to access TimeArray{Float64,1,Date,Array{Float64,1}}
+    at index []
+  ```
+
 
 ### 0.15.0